Super build

+++ packages/core/integration-tests/test/posthtml.js
@@ -108,7 +108,11 @@ describe.v2('posthtml', function () {
     assert(asset.dependencies.get(other).includedInParent);
   });
 
-  it('should automatically install posthtml plugins if needed', async () => {
+  it('should automatically install posthtml plugins if needed', async function () {
+    if (process.env.SUPER_PACKAGE === 'true') {
+      // Auto install not supported in super package
+      return this.skip();
+    }
     let inputDir = path.join(__dirname, '/input');
     await outputFS.rimraf(inputDir);
     await ncp(
